refactor(canvas): extract model switcher module
