+++ apps/canvas-workspace/src/renderer/src/components/chat/ModelSettings.tsx
@@ -1,4 +1,3 @@
 import './ModelSettings.css';
 
-export { ModelSwitcher } from './ModelSwitcher';
 export { ModelsSection } from './ModelsSection';
