+++ apps/canvas-workspace/src/renderer/src/components/models/ModelSwitcher/__tests__/ModelSwitcher.test.tsx
@@ -2,9 +2,9 @@
 import { act } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import { I18nProvider } from '../../../i18n';
-import { ModelSwitcher } from '../ModelSwitcher';
-import type { CanvasModelStatus, ModelSelection } from '../../../types';
+import { I18nProvider } from '../../../../i18n';
+import { ModelSwitcher } from '../index';
+import type { CanvasModelStatus, ModelSelection } from '../../../../types';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
