+++ apps/canvas-workspace/src/renderer/src/components/models/ModelSwitcher/index.css
@@ -1,0 +1,290 @@
+/* ─── Model switcher & AI settings ────────────────────────── */
+
+.chat-model-switcher {
+  position: relative;
+  z-index: 30;
+}
+
+.chat-model-switcher-error {
+  position: absolute;
+  right: 0;
+  bottom: calc(100% + 6px);
+  width: max-content;
+  max-width: min(320px, 72vw);
+  padding: 7px 9px;
+  border: 1px solid color-mix(in srgb, var(--error) 35%, transparent);
+  border-radius: var(--radius-sm);
+  background: var(--surface);
+  color: var(--error);
+  font-size: 11px;
+  line-height: 1.35;
+  box-shadow: var(--shadow-float);
+}
+
+.chat-model-switcher-btn {
+  height: 30px;
+  border: 1px solid rgba(55, 53, 47, 0.1);
+  background: rgba(55, 53, 47, 0.035);
+  color: #5f5e59;
+  border-radius: var(--radius-pill);
+  padding: 0 9px 0 8px;
+  display: inline-flex;
+  align-items: center;
+  gap: 6px;
+  font-size: 12px;
+  font-weight: 500;
+  cursor: pointer;
+  /* Wide enough for "provider / model" — the old 132px cut a bare model name
+     to a few characters, which is the whole reason the provider had to be
+     spelled out here in the first place. */
+  max-width: 240px;
+  transition: background 0.12s, border-color 0.12s, color 0.12s;
+}
+
+.chat-model-switcher-btn:hover {
+  background: rgba(35, 131, 226, 0.07);
+  border-color: rgba(35, 131, 226, 0.2);
+  color: #2383e2;
+}
+
+.chat-model-switcher-btn--warning {
+  background: rgba(35, 131, 226, 0.06);
+  border-color: rgba(35, 131, 226, 0.2);
+  color: #1f76cc;
+  max-width: 180px;
+}
+
+.chat-model-switcher-btn--warning:hover {
+  background: rgba(35, 131, 226, 0.1);
+  border-color: rgba(35, 131, 226, 0.32);
+  color: #1c6dbf;
+}
+
+.chat-model-switcher-btn--warning .chat-model-switcher-dot {
+  background: #2383e2;
+  opacity: 0.9;
+}
+
+.chat-model-switcher-dot {
+  width: 6px;
+  height: 6px;
+  border-radius: 50%;
+  background: currentColor;
+  opacity: 0.55;
+  flex-shrink: 0;
+}
+
+.chat-model-switcher-label {
+  min-width: 0;
+  display: inline-flex;
+  align-items: center;
+  gap: 4px;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+}
+
+/* The provider is context, the model is the answer: the provider gives up
+   its width first so the model name stays readable in a narrow composer. */
+.chat-model-switcher-provider {
+  flex: 0 3 auto;
+  min-width: 0;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  opacity: 0.62;
+}
+
+.chat-model-switcher-provider::after {
+  content: "/";
+  padding-left: 4px;
+  opacity: 0.7;
+}
+
+.chat-model-switcher-model {
+  flex: 0 1 auto;
+  min-width: 0;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+}
+
+.chat-model-switcher-chevron {
+  color: #9b9a97;
+  margin-left: -1px;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  flex-shrink: 0;
+}
+
+.chat-model-menu {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: min(360px, calc(100vw - 24px));
+  max-height: min(520px, calc(100vh - 24px));
+  overflow-y: auto;
+  background: #fff;
+  border: 1px solid rgba(55, 53, 47, 0.1);
+  border-radius: var(--radius-xl);
+  padding: 6px;
+  box-shadow: 0 18px 44px rgba(55, 53, 47, 0.13), 0 0 0 1px rgba(55, 53, 47, 0.02);
+  animation: chat-menu-in 0.12s ease-out;
+  z-index: calc(var(--layer-dock) + 10);
+}
+
+/* Sticky because the panel itself is the scroll container: the filter has to
+   stay reachable while a long catalog scrolls under it. The negative top
+   cancels the panel's own 6px padding so nothing peeks above it. */
+.chat-model-menu-controls {
+  position: sticky;
+  top: -6px;
+  z-index: 1;
+  margin: -6px -6px 2px;
+  background: var(--surface);
+  border-radius: var(--radius-xl) var(--radius-xl) 0 0;
+}
+
+.chat-model-menu-search {
+  padding: 8px 8px 6px;
+}
+
+/* ui/TextField owns the field chrome; only the menu's tighter metrics differ. */
+.chat-model-menu-search-input.ui-textfield__control {
+  height: 30px;
+  font-size: 12px;
+}
+
+.chat-model-menu-toolbar {
+  min-width: 0;
+  padding: 0 8px 7px;
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.chat-model-menu-provider-filter {
+  min-width: 0;
+  overflow-x: auto;
+  scrollbar-width: none;
+  background: transparent;
+}
+
+.chat-model-menu-provider-filter::-webkit-scrollbar {
+  display: none;
+}
+
+.chat-model-menu-provider-filter .ui-segmented__option {
+  flex-shrink: 0;
+}
+
+.chat-model-menu-result-count {
+  margin-left: auto;
+  color: var(--text-muted);
+  font-size: 10px;
+  white-space: nowrap;
+}
+
+.chat-model-menu-provider-head {
+  display: flex;
+  justify-content: space-between;
+  gap: 10px;
+  padding: 9px 9px 5px;
+  font-size: 10px;
+  line-height: 1;
+  font-weight: 500;
+  letter-spacing: 0.055em;
+  text-transform: uppercase;
+  color: rgba(55, 53, 47, 0.42);
+}
+
+.chat-model-menu-provider-head span:last-child {
+  font-weight: 500;
+  letter-spacing: 0;
+  text-transform: none;
+  color: rgba(55, 53, 47, 0.34);
+}
+
+.chat-model-menu-item,
+.chat-model-menu-action {
+  width: 100%;
+  border: none;
+  background: transparent;
+  border-radius: 7px;
+  display: flex;
+  align-items: center;
+  gap: 7px;
+  padding: 7px 8px;
+  text-align: left;
+  color: #37352f;
+  cursor: pointer;
+  transition: background 0.1s;
+}
+
+.chat-model-menu-item:hover,
+.chat-model-menu-item:focus-visible,
+.chat-model-menu-action:hover,
+.chat-model-menu-action:focus-visible {
+  outline: none;
+  background: rgba(55, 53, 47, 0.06);
+}
+
+.chat-model-menu-item--active,
+.chat-model-menu-item--active:focus-visible {
+  background: rgba(35, 131, 226, 0.08);
+}
+
+.chat-model-menu-check {
+  width: 16px;
+  color: #2383e2;
+  display: inline-flex;
+  justify-content: center;
+  flex-shrink: 0;
+}
+
+.chat-model-menu-main {
+  min-width: 0;
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+}
+
+.chat-model-menu-title,
+.chat-model-menu-subtitle {
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+}
+
+.chat-model-menu-title {
+  font-size: 13px;
+  font-weight: 500;
+}
+
+.chat-model-menu-subtitle,
+.chat-model-menu-empty,
+.chat-model-menu-hint {
+  font-size: 11px;
+  color: rgba(55, 53, 47, 0.46);
+}
+
+.chat-model-menu-empty,
+.chat-model-menu-hint {
+  padding: 7px 10px 9px 31px;
+  line-height: 1.45;
+}
+
+.chat-model-menu-divider {
+  height: 1px;
+  margin: 5px 3px;
+  background: rgba(55, 53, 47, 0.07);
+}
+
+.chat-model-menu-action {
+  justify-content: center;
+  font-size: 12px;
+  font-weight: 500;
+  color: #2383e2;
+}
